Hot fix

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ int main(void)
 	long int li;
 	long long int lli;
 	float f;
-	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
-	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(i));
-	printf("Size of long int: %lu byte(s)\n", (unsigned long)sizeof(li));
-	printf("Size of long long int: %lu byte(s)\n", (unsigned long)sizeof(lli));
-	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
+	printf("Size of a char: %d byte(s)\n", sizeof(c);
+	printf("Size of an int: %d byte(s)\n", sizeof(i);
+	printf("Size of long int: %d byte(s)\n", sizeof(li);
+	printf("Size of long long int: %d byte(s)\n", sizeof(lli);
+	printf("Size of a float: %d byte(s)\n", sizeof(f);
 	return (0);
 }
